feat - Moved timeout tests to cli tests
