+++ cli/package.json
@@ -17,8 +17,7 @@
   "scripts": {
     "build": "tsc",
     "postbuild": "node scripts/make-executable.js",
-    "test": "node scripts/cli-tests.js",
-    "test:timeouts": "node scripts/timeout-tests.js"
+    "test": "node scripts/cli-tests.js"
   },
   "devDependencies": {},
   "dependencies": {
