+++ cli/scripts/cli-tests.js
@@ -96,8 +96,85 @@ try {
 const invalidConfigPath = path.join(TEMP_DIR, "invalid-config.json");
 fs.writeFileSync(invalidConfigPath, '{\n  "mcpServers": {\n    "invalid": {');
 
+// Helper function to analyze output for timeout behavior verification
+function analyzeTimeoutBehavior(output, testName, expectedBehavior) {
+  const lines = output.split("\n");
+
+  console.log(
+    `${colors.BLUE}Analyzing timeout behavior for ${testName}...${colors.NC}`,
+  );
+
+  // Extract key timing information
+  const progressUpdates = lines.filter((line) =>
+    line.includes("Progress update received"),
+  );
+  const timeoutMessages = lines.filter(
+    (line) => line.includes("timeout") || line.includes("timed out"),
+  );
+  const errorMessages = lines.filter(
+    (line) => line.includes("Error") || line.includes("Failed"),
+  );
+
+  console.log(
+    `${colors.BLUE}Found ${progressUpdates.length} progress updates${colors.NC}`,
+  );
+
+  if (progressUpdates.length > 0) {
+    console.log(
+      `${colors.BLUE}First progress update: ${progressUpdates[0]}${colors.NC}`,
+    );
+    if (progressUpdates.length > 1) {
+      console.log(
+        `${colors.BLUE}Last progress update: ${progressUpdates[progressUpdates.length - 1]}${colors.NC}`,
+      );
+    }
+  }
+
+  if (timeoutMessages.length > 0) {
+    console.log(
+      `${colors.BLUE}Timeout messages: ${timeoutMessages.join("\n")}${colors.NC}`,
+    );
+  }
+
+  if (errorMessages.length > 0) {
+    console.log(
+      `${colors.BLUE}Error messages: ${errorMessages.join("\n")}${colors.NC}`,
+    );
+  }
+
+  switch (expectedBehavior) {
+    case "should_reset_timeout":
+      // For tests where resetTimeoutOnProgress is true, we expect:
+      // 1. Multiple progress updates (more than 2)
+      // 2. No timeout errors before completion
+      // 3. Successful completion
+      return (
+        progressUpdates.length >= 3 &&
+        !output.includes("Request timed out") &&
+        !output.includes("Maximum total timeout exceeded")
+      );
+
+    case "should_timeout_quickly":
+      // For tests where resetTimeoutOnProgress is false, we expect:
+      // 1. Few progress updates (less than 3)
+      // 2. A timeout error
+      return (
+        progressUpdates.length < 3 &&
+        (output.includes("Request timed out") || output.includes("timed out"))
+      );
+
+    case "should_hit_max_timeout":
+      // For tests where maxTotalTimeout should be hit, we expect:
+      // 1. Error mentioning maximum total timeout
+      return output.includes("Maximum total timeout exceeded");
+
+    default:
+      return null; // No specific expectation
+  }
+}
+
 // Function to run a basic test
-async function runBasicTest(testName, ...args) {
+async function runBasicTest(testName, expectedBehavior, ...args) {
   const outputFile = path.join(
     OUTPUT_DIR,
     `${testName.replace(/\//g, "_")}.log`,
@@ -105,6 +182,7 @@ async function runBasicTest(testName, ...args) {
 
   console.log(`\n${colors.YELLOW}Testing: ${testName}${colors.NC}`);
   TOTAL_TESTS++;
+  console.log("basic test kavya " + TOTAL_TESTS);
 
   // Run the command and capture output
   console.log(
@@ -136,6 +214,36 @@ async function runBasicTest(testName, ...args) {
 
       child.on("close", (code) => {
         outputStream.end();
+
+        // For specific timeout behavior tests, validate the behavior
+        if (expectedBehavior) {
+          const behaviorCorrect = analyzeTimeoutBehavior(
+            output,
+            testName,
+            expectedBehavior,
+          );
+
+          if (behaviorCorrect === true) {
+            console.log(
+              `${colors.GREEN}✓ Timeout behavior test passed: ${testName}${colors.NC}`,
+            );
+            PASSED_TESTS++;
+            resolve(true);
+            return;
+          } else if (behaviorCorrect === false) {
+            console.log(
+              `${colors.RED}✗ Timeout behavior test failed: ${testName}${colors.NC}`,
+            );
+            console.log(
+              `${colors.RED}Expected: ${expectedBehavior}${colors.NC}`,
+            );
+            console.log(
+              `${colors.RED}Output did not match expected behavior${colors.NC}`,
+            );
+            FAILED_TESTS++;
+            process.exit(1);
+          }
+        }
 
         if (code === 0) {
           console.log(`${colors.GREEN}✓ Test passed: ${testName}${colors.NC}`);
@@ -177,7 +285,7 @@ async function runBasicTest(testName, ...args) {
 }
 
 // Function to run an error test (expected to fail)
-async function runErrorTest(testName, ...args) {
+async function runErrorTest(testName, expectedBehavior, ...args) {
   const outputFile = path.join(
     OUTPUT_DIR,
     `${testName.replace(/\//g, "_")}.log`,
@@ -185,6 +293,7 @@ async function runErrorTest(testName, ...args) {
 
   console.log(`\n${colors.YELLOW}Testing error case: ${testName}${colors.NC}`);
   TOTAL_TESTS++;
+  console.log("error test kavya " + TOTAL_TESTS);
 
   // Run the command and capture output
   console.log(
@@ -217,11 +326,52 @@ async function runErrorTest(testName, ...args) {
       child.on("close", (code) => {
         outputStream.end();
 
+        // For specific timeout behavior tests, validate the behavior
+        if (expectedBehavior) {
+          const behaviorCorrect = analyzeTimeoutBehavior(
+            output,
+            testName,
+            expectedBehavior,
+          );
+
+          if (behaviorCorrect === true) {
+            console.log(
+              `${colors.GREEN}✓ Timeout behavior test passed: ${testName}${colors.NC}`,
+            );
+            PASSED_TESTS++;
+            resolve(true);
+            return;
+          } else if (behaviorCorrect === false) {
+            console.log(
+              `${colors.RED}✗ Timeout behavior test failed: ${testName}${colors.NC}`,
+            );
+            console.log(
+              `${colors.RED}Expected: ${expectedBehavior}${colors.NC}`,
+            );
+            console.log(
+              `${colors.RED}Output did not match expected behavior${colors.NC}`,
+            );
+            FAILED_TESTS++;
+            process.exit(1);
+          }
+        }
+
         // For error tests, we expect a non-zero exit code
         if (code !== 0) {
-          console.log(
-            `${colors.GREEN}✓ Error test passed: ${testName}${colors.NC}`,
-          );
+          // Look for specific timeout errors
+          if (
+            testName.includes("timeout") &&
+            (output.includes("timeout") || output.includes("timed out"))
+          ) {
+            console.log(
+              `${colors.GREEN}✓ Timeout error test passed: ${testName}${colors.NC}`,
+            );
+          } else {
+            console.log(
+              `${colors.GREEN}✓ Error test passed: ${testName}${colors.NC}`,
+            );
+          }
+
           console.log(`${colors.BLUE}Error output (expected):${colors.NC}`);
           const firstFewLines = output
             .split("\n")
@@ -270,6 +420,7 @@ async function runTests() {
   // Test 1: Basic CLI mode with method
   await runBasicTest(
     "basic_cli_mode",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "--cli",
@@ -280,6 +431,7 @@ async function runTests() {
   // Test 2: CLI mode with non-existent method (should fail)
   await runErrorTest(
     "nonexistent_method",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "--cli",
@@ -288,7 +440,7 @@ async function runTests() {
   );
 
   // Test 3: CLI mode without method (should fail)
-  await runErrorTest("missing_method", TEST_CMD, ...TEST_ARGS, "--cli");
+  await runErrorTest("missing_method", null, TEST_CMD, ...TEST_ARGS, "--cli");
 
   console.log(
     `\n${colors.YELLOW}=== Running Environment Variable Tests ===${colors.NC}`,
@@ -297,6 +449,7 @@ async function runTests() {
   // Test 4: CLI mode with environment variables
   await runBasicTest(
     "env_variables",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "-e",
@@ -311,6 +464,7 @@ async function runTests() {
   // Test 5: CLI mode with invalid environment variable format (should fail)
   await runErrorTest(
     "invalid_env_format",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "-e",
@@ -323,6 +477,7 @@ async function runTests() {
   // Test 5b: CLI mode with environment variable containing equals sign in value
   await runBasicTest(
     "env_variable_with_equals",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "-e",
@@ -335,6 +490,7 @@ async function runTests() {
   // Test 5c: CLI mode with environment variable containing base64-encoded value
   await runBasicTest(
     "env_variable_with_base64",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "-e",
@@ -351,6 +507,7 @@ async function runTests() {
   // Test 6: Using config file with CLI mode
   await runBasicTest(
     "config_file",
+    null,
     "--config",
     path.join(PROJECT_ROOT, "sample-config.json"),
     "--server",
@@ -363,6 +520,7 @@ async function runTests() {
   // Test 7: Using config file without server name (should fail)
   await runErrorTest(
     "config_without_server",
+    null,
     "--config",
     path.join(PROJECT_ROOT, "sample-config.json"),
     "--cli",
@@ -373,6 +531,7 @@ async function runTests() {
   // Test 8: Using server name without config file (should fail)
   await runErrorTest(
     "server_without_config",
+    null,
     "--server",
     "everything",
     "--cli",
@@ -383,6 +542,7 @@ async function runTests() {
   // Test 9: Using non-existent config file (should fail)
   await runErrorTest(
     "nonexistent_config",
+    null,
     "--config",
     "./nonexistent-config.json",
     "--server",
@@ -395,6 +555,7 @@ async function runTests() {
   // Test 10: Using invalid config file format (should fail)
   await runErrorTest(
     "invalid_config",
+    null,
     "--config",
     invalidConfigPath,
     "--server",
@@ -407,6 +568,7 @@ async function runTests() {
   // Test 11: Using config file with non-existent server (should fail)
   await runErrorTest(
     "nonexistent_server",
+    null,
     "--config",
     path.join(PROJECT_ROOT, "sample-config.json"),
     "--server",
@@ -423,6 +585,7 @@ async function runTests() {
   // Test 12: CLI mode with tool call
   await runBasicTest(
     "tool_call",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "--cli",
@@ -437,6 +600,7 @@ async function runTests() {
   // Test 13: CLI mode with tool call but missing tool name (should fail)
   await runErrorTest(
     "missing_tool_name",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "--cli",
@@ -449,6 +613,7 @@ async function runTests() {
   // Test 14: CLI mode with tool call but invalid tool args format (should fail)
   await runErrorTest(
     "invalid_tool_args",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "--cli",
@@ -463,6 +628,7 @@ async function runTests() {
   // Test 15: CLI mode with multiple tool args
   await runBasicTest(
     "multiple_tool_args",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "--cli",
@@ -482,6 +648,7 @@ async function runTests() {
   // Test 16: CLI mode with resource read
   await runBasicTest(
     "resource_read",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "--cli",
@@ -494,6 +661,7 @@ async function runTests() {
   // Test 17: CLI mode with resource read but missing URI (should fail)
   await runErrorTest(
     "missing_uri",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "--cli",
@@ -508,6 +676,7 @@ async function runTests() {
   // Test 18: CLI mode with prompt get
   await runBasicTest(
     "prompt_get",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "--cli",
@@ -520,6 +689,7 @@ async function runTests() {
   // Test 19: CLI mode with prompt get and args
   await runBasicTest(
     "prompt_get_with_args",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "--cli",
@@ -535,6 +705,7 @@ async function runTests() {
   // Test 20: CLI mode with prompt get but missing prompt name (should fail)
   await runErrorTest(
     "missing_prompt_name",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "--cli",
@@ -547,6 +718,7 @@ async function runTests() {
   // Test 21: CLI mode with log level
   await runBasicTest(
     "log_level",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "--cli",
@@ -559,6 +731,7 @@ async function runTests() {
   // Test 22: CLI mode with invalid log level (should fail)
   await runErrorTest(
     "invalid_log_level",
+    null,
     TEST_CMD,
     ...TEST_ARGS,
     "--cli",
@@ -580,6 +753,7 @@ async function runTests() {
   // Test 23: CLI mode with config file, environment variables, and tool call
   await runBasicTest(
     "combined_options",
+    null,
     "--config",
     path.join(PROJECT_ROOT, "sample-config.json"),
     "--server",
@@ -594,6 +768,7 @@ async function runTests() {
   // Test 24: CLI mode with all possible options (that make sense together)
   await runBasicTest(
     "all_options",
+    null,
     "--config",
     path.join(PROJECT_ROOT, "sample-config.json"),
     "--server",
@@ -609,6 +784,186 @@ async function runTests() {
     "message=Hello",
     "--log-level",
     "debug",
+  );
+
+  console.log(
+    `${colors.YELLOW}=== MCP Inspector CLI Timeout Configuration Tests ===${colors.NC}`,
+  );
+  console.log(
+    `${colors.BLUE}This script tests the MCP Inspector CLI's timeout configuration options:${colors.NC}`,
+  );
+  console.log(
+    `${colors.BLUE}- Request timeout (--request-timeout)${colors.NC}`,
+  );
+  console.log(
+    `${colors.BLUE}- Reset timeout on progress (--reset-timeout-on-progress)${colors.NC}`,
+  );
+  console.log(
+    `${colors.BLUE}- Maximum total timeout (--max-total-timeout)${colors.NC}\n`,
+  );
+
+  // Test 25: Default timeout values
+  await runBasicTest(
+    "default_timeouts",
+    null,
+    TEST_CMD,
+    ...TEST_ARGS,
+    "--cli",
+    "--method",
+    "tools/list",
+  );
+
+  // Test 26: Custom request timeout
+  await runBasicTest(
+    "custom_request_timeout",
+    null,
+    TEST_CMD,
+    ...TEST_ARGS,
+    "--cli",
+    "--method",
+    "tools/call",
+    "--tool-name",
+    "longRunningOperation",
+    "--tool-arg",
+    "duration=5",
+    "steps=5",
+    "--request-timeout",
+    "15000",
+  );
+
+  // Test 27: Request timeout too short (should fail)
+  await runErrorTest(
+    "short_request_timeout",
+    "should_timeout_quickly",
+    TEST_CMD,
+    ...TEST_ARGS,
+    "--cli",
+    "--method",
+    "tools/call",
+    "--tool-name",
+    "longRunningOperation",
+    "--tool-arg",
+    "duration=5",
+    "steps=5",
+    "--request-timeout",
+    "100",
+  );
+
+  console.log(
+    `\n${colors.YELLOW}=== Running Progress-Related Timeout Tests ===${colors.NC}`,
+  );
+
+  // Test 28: Reset timeout on progress enabled - should complete successfully
+  // await runBasicTest(
+  //   "reset_timeout_on_progress",
+  //   "should_reset_timeout",
+  //   TEST_CMD,
+  //   ...TEST_ARGS,
+  //   "--cli",
+  //   "--method",
+  //   "tools/call",
+  //   "--tool-name",
+  //   "longRunningOperation",
+  //   "--tool-arg",
+  //   "duration=15", // 15 second operation
+  //   "steps=5", // 5 steps = progress every 3 seconds
+  //   "--request-timeout",
+  //   "2000", // 2 second timeout per interval
+  //   "--reset-timeout-on-progress",
+  //   "true",
+  //   "--max-total-timeout",
+  //   "30000",
+  // );
+
+  // Test 29: Reset timeout on progress disabled - should fail with timeout
+  await runErrorTest(
+    "reset_timeout_disabled",
+    "should_timeout_quickly",
+    TEST_CMD,
+    ...TEST_ARGS,
+    "--cli",
+    "--method",
+    "tools/call",
+    "--tool-name",
+    "longRunningOperation",
+    "--tool-arg",
+    "duration=15", // Same configuration as above
+    "steps=5",
+    "--request-timeout",
+    "2000",
+    "--reset-timeout-on-progress",
+    "false", // Only difference is here
+    "--max-total-timeout",
+    "30000",
+  );
+
+  // console.log(
+  //   `\n${colors.YELLOW}=== Running Max Total Timeout Tests ===${colors.NC}`,
+  // );
+
+  // Test 30: Max total timeout exceeded (should fail)
+  // await runErrorTest(
+  //   "max_total_timeout_exceeded",
+  //   "should_hit_max_timeout",
+  //   TEST_CMD,
+  //   ...TEST_ARGS,
+  //   "--cli",
+  //   "--method",
+  //   "tools/call",
+  //   "--tool-name",
+  //   "longRunningOperation",
+  //   "--tool-arg",
+  //   "duration=10",
+  //   "steps=10",
+  //   "--request-timeout",
+  //   "2000",
+  //   "--reset-timeout-on-progress",
+  //   "true",
+  //   "--max-total-timeout",
+  //   "3000", // 3 second total timeout
+  // );
+
+  console.log(
+    `\n${colors.YELLOW}=== Running Input Validation Tests ===${colors.NC}`,
+  );
+
+  // Test 31: Invalid request timeout value
+  await runErrorTest(
+    "invalid_request_timeout",
+    null,
+    TEST_CMD,
+    ...TEST_ARGS,
+    "--cli",
+    "--method",
+    "tools/list",
+    "--request-timeout",
+    "invalid",
+  );
+
+  // Test 32: Invalid reset-timeout-on-progress value
+  await runErrorTest(
+    "invalid_reset_timeout",
+    null,
+    TEST_CMD,
+    ...TEST_ARGS,
+    "--cli",
+    "--method",
+    "tools/list",
+    "--reset-timeout-on-progress",
+    "not-a-boolean",
+  );
+
+  // Test 33: Invalid max total timeout value
+  await runErrorTest(
+    "invalid_max_timeout",
+    null,
+    TEST_CMD,
+    ...TEST_ARGS,
+    "--cli",
+    "--method",
+    "tools/list",
+    "--max-total-timeout",
+    "invalid",
   );
 
   // Print test summary
